Fixed tabs in clusted details page and added it in backup page as well
